// models/destination.model.ts
export interface DestinationModel {
  id: string;

  title_en: string;
  title_ru: string;
  title_tr: string;
  country_en: string;
  country_ru: string;
  country_tr: string;
  city_en: string;
  city_ru: string;
  city_tr: string;

  imagePath: string;

  createdAt: string;
  updatedAt: string;
}